Best cash back view updates
Implemented best cash back view from the figma file, but did not remove the arrows on the side yet. Also added a functionality so that if the background of the card if dark, it will show as a white font and if the background it light, it will show as a black font.